// frontend/Angular/Angular/src/app/clients/users-api.client.ts
export class CreatedUserDto implements ICreatedUserDto {
    name!: string | undefined;
    email!: string | undefined;
    password!: string | undefined;
    age!: number;
    roles!: string[] | undefined;

    constructor(data?: ICreatedUserDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.name = _data["name"];
            this.email = _data["email"];
            this.password = _data["password"];
            this.age = _data["age"];
            if (Array.isArray(_data["roles"])) {
                this.roles = [] as any;
                for (let item of _data["roles"])
                    this.roles!.push(item);
            }
        }
    }

    static fromJS(data: any): CreatedUserDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreatedUserDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["email"] = this.email;
        data["password"] = this.password;
        data["age"] = this.age;
        if (Array.isArray(this.roles)) {
            data["roles"] = [];
            for (let item of this.roles)
                data["roles"].push(item);
        }
        return data;
    }
}

export interface ICreatedUserDto {
    name: string | undefined;
    email: string | undefined;
    password: string | undefined;
    age: number;
    roles: string[] | undefined;
}
// ...
export interface IUserDto {
    id: number;
    name: string | undefined;
    email: string | undefined;
    age: number;
    roles: string[] | undefined;
}
// ...
export class UserRoleDto implements IUserRoleDto {
    role!: string | undefined;

    constructor(data?: IUserRoleDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.role = _data["role"];
        }
    }

    static fromJS(data: any): UserRoleDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserRoleDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["role"] = this.role;
        return data;
    }
}

export interface IUserRoleDto {
    role: string | undefined;
}